Se agregó el método para traer el atributo estado de la clase Mascota en las capas controller, service y repository

+++ src/main/java/com/veterinaria/veterinaria/Services/AdoptanteService.java
@@ -4,6 +4,7 @@ import com.veterinaria.veterinaria.Entity.Adoptante;
 import com.veterinaria.veterinaria.Repository.AdoptanteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.veterinaria.veterinaria.Entity.Mascota;
 
 import java.util.List;
 import java.util.Optional;
@@ -22,7 +23,7 @@ public class AdoptanteService {
     //Optional en el metodo findById que ayuda a manejar la posibilidad de que no se encuentre un recurso.
     // Si llamamos a findById y no existe un Adoptante con el ID proporcionado, Optional permite evitar
     // excepciones NullPointerException
-    public Optional<Adoptante> findById(Integer id) {
+    public Optional<Adoptante> findById(Long id) {
         return adoptanteRepository.findById(id);
     }
 
@@ -30,7 +31,8 @@ public class AdoptanteService {
         return adoptanteRepository.save(adoptante);
     }
 
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         adoptanteRepository.deleteById(id);
     }
 }
+
